refactor #260526: レビュー対応: RPC、プロファイル作成モーダルなど
